requirements.txt
akasha-terminal
